Add support to systemd journal with the JOURNALCTL_CMD configuration directive or the -j command line option.
The journalctl command to use can be defined instead of log file entry. For
example, with postfix it migth be set to the following:

        JOURNALCTL_CMD  journalctl -u postfix

and for sendmail:

        JOURNALCTL_CMD  journalctl -u sendmail

When enabled, the LOG_FILE configuration directive above is just ommitted.
The additional option: --output="short-iso" is also always used to format
timestamp.

In incremental mode sendmailanalyzer will automatically set the --since option
to the last parsed timestamp to prevent loading previous messages.

Note that in daemon mode sendmailanalyzer will automatically add the -f option
to the command. Can be overwritten with --journalctl or -j options.

This patch fix also an issue with command line option in uppercase and not.