Added database lookup of virtual domains, so sendmailanalyzer is not limited to the LOCAL_DOMAIN configuration directive. This is useful for mailserver which have virtual users and domains in their database and therefor no static configuration. Here are the new configuration directives added to the sendmailanalyzer configuration file and their explanation:
  - VIRTUAL_DOMAIN_DB	DBI:mysql:database=mailserver:host=localhost

Leave this unconfigured if you don't have virtual domains in a database
set up or set the value to a valid Perl DBI DSN (data source name). Here
this is a connection to a MySQL database so it require that Perl module
DBI and DBD::MySQL are installed on your system.

  - VIRTUAL_DOMAIN_DB_USER and VIRTUAL_DOMAIN_DB_PASS

The user and password to use to connect to the database.

  - VIRTUAL_DOMAIN_DB_QUERY	SELECT name FROM virtual_domains

The SQL query to use to retrieve the list of virtual domain that will
be appended to the LOCAL_DOMAIN array. See LOCAL_DOMAIN for more
information.

Thanks to thinksilicon for the patch.